+++ internal/handler/video_favorite.go
@@ -1,0 +1,557 @@
+package handler
+
+import (
+	"cakecake/internal/errcode"
+	"cakecake/internal/middleware"
+	"cakecake/internal/model/video"
+	"cakecake/internal/pkg/resp"
+	"context"
+	"net/http"
+	"strconv"
+
+	"github.com/gin-gonic/gin"
+)
+
+func (a *API) userVideoFavoriteCount(ctx context.Context, uid, vid uint64) (int64, error) {
+	return a.FavoriteSvc.UserFavoriteCount(ctx, uid, vid)
+}
+
+func (a *API) engagementByViewer(ctx context.Context, viewer uint64, ids []uint64) map[uint64]videoEngagement {
+	out := make(map[uint64]videoEngagement, len(ids))
+	if viewer == 0 || len(ids) == 0 {
+		return out
+	}
+	liked := a.EngagementSvc.BatchVideoLikes(ctx, viewer, ids)
+	faved := a.FavoriteSvc.BatchFavorited(ctx, viewer, ids)
+	coined := a.EngagementSvc.BatchCoinedByUser(ctx, viewer, ids)
+	later := a.EngagementSvc.BatchWatchLater(ctx, viewer, ids)
+	for _, id := range ids {
+		coinAmt := coined[id]
+		out[id] = videoEngagement{
+			LikedByMe:     liked[id],
+			FavoritedByMe: faved[id],
+			CoinedByMe:    coinAmt > 0,
+			MyCoinAmount:  coinAmt,
+			InWatchLater:  later[id],
+		}
+	}
+	return out
+}
+
+func loadPublishedVideo(ctx context.Context, a *API, vid uint64) (*video.Video, bool) {
+	v, err := a.VideoSvc.GetPublishedVideo(ctx, vid)
+	if err != nil {
+		return nil, false
+	}
+	return v, true
+}
+
+// ToggleVideoFavorite toggles the current user's favorite on a published video.
+// ToggleVideoFavorite godoc
+// @Summary      Favorite/unfavorite a video
+// @Description  Toggle favorite status on a video
+// @Tags         Videos
+// @Produce      json
+// @Param        id path int true "Video ID"
+// @Success      200 {object} map[string]interface{}
+// @Router       /videos/{id}/favorite [post]
+func (a *API) ToggleVideoFavorite(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	vid, err := strconv.ParseUint(c.Param("id"), 10, 64)
+	if err != nil || vid == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if _, ok := loadPublishedVideo(c.Request.Context(), a, vid); !ok {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	def, err := a.ensureDefaultFavoriteFolder(c.Request.Context(), uid)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	favorited, favCount, err := a.FavoriteSvc.ToggleVideoFavoriteWithFolder(c.Request.Context(), uid, vid, def.ID)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	resp.OK(c, videoFavToggleResponse{Favorited: favorited, FavCount: favCount})
+}
+
+const favoriteFolderCapacity = 999
+
+type videoFavToggleResponse struct {
+	Favorited bool   `json:"favorited"`
+	FavCount  uint64 `json:"fav_count"`
+}
+
+type favoriteFolderSelectItem struct {
+	ID         uint64 `json:"id"`
+	Title      string `json:"title"`
+	IsDefault  bool   `json:"is_default"`
+	VideoCount int64  `json:"video_count"`
+	CountLabel string `json:"count_label"`
+	Selected   bool   `json:"selected"`
+}
+
+type videoFavFolderSelectResponse struct {
+	Favorited bool                       `json:"favorited"`
+	FavCount  uint64                     `json:"fav_count"`
+	FolderIDs []uint64                   `json:"folder_ids"`
+	Items     []favoriteFolderSelectItem `json:"items"`
+}
+
+type videoFavSetResponse struct {
+	Favorited bool     `json:"favorited"`
+	FavCount  uint64   `json:"fav_count"`
+	FolderIDs []uint64 `json:"folder_ids"`
+}
+
+type videoFavRemoveResponse struct {
+	OK      bool `json:"ok"`
+	Removed bool `json:"removed"`
+}
+
+type videoFavCopyResponse struct {
+	OK     bool `json:"ok"`
+	Copied bool `json:"copied"`
+}
+
+type videoFavMoveResponse struct {
+	OK    bool `json:"ok"`
+	Moved bool `json:"moved"`
+}
+
+type favoriteVideoItemDTO struct {
+	ID                uint64 `json:"id"`
+	Title             string `json:"title"`
+	CoverURL          string `json:"cover_url"`
+	PlayCount         uint64 `json:"play_count"`
+	DanmakuCount      uint64 `json:"danmaku_count"`
+	Duration          uint64 `json:"duration"`
+	Uploader          string `json:"uploader"`
+	UploaderID        uint64 `json:"uploader_id"`
+	UploaderAvatarURL string `json:"uploader_avatar_url"`
+	CreatedAt         string `json:"created_at"`
+	FavoritedAt       string `json:"favorited_at"`
+	FolderID          uint64 `json:"folder_id"`
+}
+
+type favoriteListResponse struct {
+	Items    []favoriteVideoItemDTO `json:"items"`
+	Total    int64                  `json:"total"`
+	MaxLimit int                    `json:"max_limit"`
+}
+
+type setVideoFavoriteFoldersJSON struct {
+	FolderIDs []uint64 `json:"folder_ids"`
+}
+
+// GetVideoFavoritePicker returns folders for the collect dialog on the video page.
+// GetVideoFavoritePicker godoc
+// @Summary      Get favorite folder picker
+// @Description  Get favorite folders for the video picker UI
+// @Tags         Videos
+// @Produce      json
+// @Param        id path int true "Video ID"
+// @Success      200 {object} map[string]interface{}
+// @Router       /videos/{id}/favorite-picker [get]
+func (a *API) GetVideoFavoritePicker(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	vid, err := strconv.ParseUint(c.Param("id"), 10, 64)
+	if err != nil || vid == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if _, ok := loadPublishedVideo(c.Request.Context(), a, vid); !ok {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	folderRows, err := a.folderListPayload(c.Request.Context(), uid, false)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	selected := make(map[uint64]bool)
+	items := make([]favoriteFolderSelectItem, 0, len(folderRows))
+	for _, row := range folderRows {
+		countLabel := strconv.FormatInt(row.VideoCount, 10)
+		if !row.IsDefault {
+			countLabel = strconv.FormatInt(row.VideoCount, 10) + "/" + strconv.Itoa(favoriteFolderCapacity)
+		}
+		inFolder, _ := a.FavoriteSvc.CheckFavoriteExists(c.Request.Context(), uid, row.ID, vid)
+		if inFolder {
+			selected[row.ID] = true
+		}
+		items = append(items, favoriteFolderSelectItem{
+			ID: row.ID, Title: row.Title,
+			IsDefault: row.IsDefault, VideoCount: row.VideoCount,
+			CountLabel: countLabel, Selected: selected[row.ID],
+		})
+	}
+	v, _ := a.VideoSvc.GetPublishedVideo(c.Request.Context(), vid)
+	var favCount uint64
+	if v != nil {
+		favCount = v.FavCount
+	}
+	resp.OK(c, videoFavFolderSelectResponse{Favorited: len(selected) > 0, FavCount: favCount, FolderIDs: folderIDsFromMap(selected), Items: items})
+}
+
+func folderIDsFromMap(m map[uint64]bool) []uint64 {
+	out := make([]uint64, 0, len(m))
+	for id := range m {
+		if id > 0 {
+			out = append(out, id)
+		}
+	}
+	return out
+}
+
+// SetVideoFavoriteFolders syncs which folders contain the video for the current user.
+// SetVideoFavoriteFolders godoc
+// @Summary      Set video favorite folders
+// @Description  Replace all favorite folders for a video
+// @Tags         Videos
+// @Produce      json
+// @Param        id path int true "Video ID"
+// @Param        body body object{folder_ids=array} true "Folder IDs"
+// @Success      200 {object} map[string]interface{}
+// @Router       /videos/{id}/favorite-folders [put]
+func (a *API) SetVideoFavoriteFolders(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	vid, err := strconv.ParseUint(c.Param("id"), 10, 64)
+	if err != nil || vid == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if _, ok := loadPublishedVideo(c.Request.Context(), a, vid); !ok {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	var body setVideoFavoriteFoldersJSON
+	if err := c.ShouldBindJSON(&body); err != nil {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if _, err := a.ensureDefaultFavoriteFolder(c.Request.Context(), uid); err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	result, err := a.FavoriteSvc.SetVideoFavoriteFolders(c.Request.Context(), uid, vid, body.FolderIDs)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if result == nil {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	resp.OK(c, videoFavSetResponse{Favorited: result.Favorited, FavCount: result.FavCount, FolderIDs: result.FolderIDs})
+}
+
+func (a *API) syncVideoFavCountAfterUserChange(ctx context.Context, vid uint64, before, after int64) {
+	if before == 0 && after > 0 {
+		_ = a.FavoriteSvc.AdjustVideoFavCount(ctx, vid, 1)
+	} else if before > 0 && after == 0 {
+		_ = a.FavoriteSvc.AdjustVideoFavCount(ctx, vid, -1)
+	}
+}
+
+func (a *API) validateFolderOwned(ctx context.Context, uid, folderID uint64) bool {
+	f, err := a.FavoriteSvc.GetFolderByID(ctx, folderID)
+	return err == nil && f.UserID == uid
+}
+
+// RemoveVideoFromFavoriteFolder removes the video from one folder (current-folder unfavorite).
+// RemoveVideoFromFavoriteFolder godoc
+// @Summary      Remove video from favorite folder
+// @Description  Remove a video from a specific favorite folder
+// @Tags         Videos
+// @Produce      json
+// @Param        id path int true "Video ID"
+// @Param        folderId path int true "Folder ID"
+// @Success      200 {object} map[string]interface{}
+// @Router       /videos/{id}/favorite-folders/{folderId} [delete]
+func (a *API) RemoveVideoFromFavoriteFolder(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	vid, folderID, ok := parseVideoFolderParams(c)
+	if !ok {
+		return
+	}
+	if _, ok := loadPublishedVideo(c.Request.Context(), a, vid); !ok {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	if !a.validateFolderOwned(c.Request.Context(), uid, folderID) {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	before, err := a.userVideoFavoriteCount(c.Request.Context(), uid, vid)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if err := a.FavoriteSvc.RemoveFavorite(c.Request.Context(), folderID, vid); err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	after, _ := a.userVideoFavoriteCount(c.Request.Context(), uid, vid)
+	a.syncVideoFavCountAfterUserChange(c.Request.Context(), vid, before, after)
+	resp.OK(c, videoFavRemoveResponse{OK: true, Removed: after < before})
+}
+
+// AddVideoToFavoriteFolder copies the video into another folder.
+// AddVideoToFavoriteFolder godoc
+// @Summary      Add video to favorite folder
+// @Description  Add a video to a specific favorite folder
+// @Tags         Videos
+// @Produce      json
+// @Param        id path int true "Video ID"
+// @Param        folderId path int true "Folder ID"
+// @Success      200 {object} map[string]interface{}
+// @Router       /videos/{id}/favorite-folders/{folderId} [post]
+func (a *API) AddVideoToFavoriteFolder(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	vid, folderID, ok := parseVideoFolderParams(c)
+	if !ok {
+		return
+	}
+	if _, ok := loadPublishedVideo(c.Request.Context(), a, vid); !ok {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	if !a.validateFolderOwned(c.Request.Context(), uid, folderID) {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	exists, err := a.FavoriteSvc.CheckFavoriteExists(c.Request.Context(), uid, folderID, vid)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if exists {
+		resp.OK(c, videoFavCopyResponse{OK: true, Copied: false})
+		return
+	}
+	cnt, err := a.FavoriteSvc.CountFavoritesInFolder(c.Request.Context(), folderID)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if cnt >= favoriteFolderCapacity {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	before, err := a.userVideoFavoriteCount(c.Request.Context(), uid, vid)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if err := a.FavoriteSvc.AddFavorite(c.Request.Context(), folderID, vid, uid); err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	after, _ := a.userVideoFavoriteCount(c.Request.Context(), uid, vid)
+	a.syncVideoFavCountAfterUserChange(c.Request.Context(), vid, before, after)
+	resp.OK(c, videoFavCopyResponse{OK: true, Copied: true})
+}
+
+type moveVideoFavoriteFolderJSON struct {
+	FromFolderID uint64 `json:"from_folder_id"`
+	ToFolderID   uint64 `json:"to_folder_id"`
+}
+
+// MoveVideoFavoriteFolder moves the video from one folder to another.
+// MoveVideoFavoriteFolder godoc
+// @Summary      Move video between favorite folders
+// @Description  Move a video from one favorite folder to another
+// @Tags         Videos
+// @Produce      json
+// @Param        id path int true "Video ID"
+// @Success      200 {object} map[string]interface{}
+// @Router       /videos/{id}/favorite-folders/move [put]
+func (a *API) MoveVideoFavoriteFolder(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	vid, err := strconv.ParseUint(c.Param("id"), 10, 64)
+	if err != nil || vid == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	var body moveVideoFavoriteFolderJSON
+	if err := c.ShouldBindJSON(&body); err != nil {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if body.FromFolderID == 0 || body.ToFolderID == 0 || body.FromFolderID == body.ToFolderID {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if _, ok := loadPublishedVideo(c.Request.Context(), a, vid); !ok {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	if !a.validateFolderOwned(c.Request.Context(), uid, body.FromFolderID) || !a.validateFolderOwned(c.Request.Context(), uid, body.ToFolderID) {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	inFrom, err := a.FavoriteSvc.CheckFavoriteExists(c.Request.Context(), uid, body.FromFolderID, vid)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if !inFrom {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	inTo, _ := a.FavoriteSvc.CheckFavoriteExists(c.Request.Context(), uid, body.ToFolderID, vid)
+	if inTo {
+		if err := a.FavoriteSvc.DeleteFavoriteByVideo(c.Request.Context(), uid, body.FromFolderID, vid); err != nil {
+			resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+			return
+		}
+		resp.OK(c, videoFavMoveResponse{OK: true, Moved: true})
+		return
+	}
+	cnt, err := a.FavoriteSvc.CountFavoritesInFolder(c.Request.Context(), body.ToFolderID)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if cnt >= favoriteFolderCapacity {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if err := a.FavoriteSvc.DeleteFavoriteByVideo(c.Request.Context(), uid, body.FromFolderID, vid); err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	if err := a.FavoriteSvc.AddFavorite(c.Request.Context(), body.ToFolderID, vid, uid); err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	resp.OK(c, videoFavMoveResponse{OK: true, Moved: true})
+}
+
+func parseVideoFolderParams(c *gin.Context) (vid, folderID uint64, ok bool) {
+	vid, err := strconv.ParseUint(c.Param("id"), 10, 64)
+	if err != nil || vid == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return 0, 0, false
+	}
+	folderID, err = strconv.ParseUint(c.Param("folderId"), 10, 64)
+	if err != nil || folderID == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return 0, 0, false
+	}
+	return vid, folderID, true
+}
+
+func (a *API) favoriteListItems(ctx context.Context, ownerID uint64, limit int, folderID uint64, filterFolder bool) ([]favoriteVideoItemDTO, int64, error) {
+	if _, err := a.ensureDefaultFavoriteFolder(ctx, ownerID); err != nil {
+		return nil, 0, err
+	}
+	result, err := a.FavoriteSvc.ListUserFavoriteVideos(ctx, ownerID, limit, folderID, filterFolder)
+	if err != nil {
+		return nil, 0, err
+	}
+	items := make([]favoriteVideoItemDTO, 0, len(result.Items))
+	for _, item := range result.Items {
+		items = append(items, favoriteVideoItemDTO{
+			ID: item.ID, Title: item.Title, CoverURL: item.CoverURL,
+			PlayCount: item.PlayCount, DanmakuCount: item.DanmakuCount, Duration: item.Duration,
+			Uploader: item.UploaderName, UploaderID: item.UploaderID,
+			UploaderAvatarURL: item.UploaderAvatar,
+			CreatedAt:         item.CreatedAt, FavoritedAt: item.FavoritedAt,
+			FolderID: item.FolderID,
+		})
+	}
+	return items, result.Total, nil
+}
+
+// ListMyFavorites returns the caller's favorited published videos (newest favorite first).
+func (a *API) ListMyFavorites(c *gin.Context) {
+	uid, ok := middleware.UserID(c)
+	if !ok {
+		resp.Err(c, http.StatusUnauthorized, errcode.CodeUnauthorized)
+		return
+	}
+	limit := parseLimit(c, 200, 200)
+	folderID, filterFolder, err := parseFolderIDQuery(c)
+	if err != nil {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	items, total, err := a.favoriteListItems(c, uid, limit, folderID, filterFolder)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	resp.OK(c, favoriteListResponse{Items: items, Total: total, MaxLimit: watchLaterMaxItems})
+}
+
+// ListUserFavorites returns a user's favorited published videos for their public space.
+func (a *API) ListUserFavorites(c *gin.Context) {
+	ownerID, err := strconv.ParseUint(c.Param("userId"), 10, 64)
+	if err != nil || ownerID == 0 {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	up, err := a.UserSvc.GetUserPublic(c.Request.Context(), ownerID)
+	if err != nil {
+		resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+		return
+	}
+	viewer, viewerOK := middleware.UserID(c)
+	if !spaceViewerCanSee(ownerID, viewerOK, viewer, up.PrivacyPublicFavorites) {
+		resp.OK(c, favoriteListResponse{Items: []favoriteVideoItemDTO{}, Total: 0})
+		return
+	}
+	limit := parseLimit(c, 200, 200)
+	folderID, filterFolder, err := parseFolderIDQuery(c)
+	if err != nil {
+		resp.Err(c, http.StatusBadRequest, errcode.CodeParamError)
+		return
+	}
+	if filterFolder {
+		f, err := a.FavoriteSvc.GetFolderByID(c.Request.Context(), folderID)
+		if err != nil || f.UserID != ownerID {
+			resp.Err(c, http.StatusNotFound, errcode.CodeNotFound)
+			return
+		}
+		if !f.IsPublic {
+			resp.Err(c, http.StatusForbidden, errcode.CodeForbidden)
+			return
+		}
+	}
+	items, total, err := a.favoriteListItems(c.Request.Context(), ownerID, limit, folderID, filterFolder)
+	if err != nil {
+		resp.Err(c, http.StatusInternalServerError, errcode.CodeInternalError)
+		return
+	}
+	resp.OK(c, favoriteListResponse{Items: items, Total: total, MaxLimit: watchLaterMaxItems})
+}
